ST SPC5 MCU port (and more).

+++ src/drivers/can.h
@@ -41,7 +41,7 @@
 struct can_frame_t {
     uint32_t id;          /* Frame ID. */
     int extended_frame;   /* Extended frame. Id is (29 bits). */
-    int size;             /* Number of bytes in data array. */
+    size_t size;          /* Number of bytes in data array. */
     int rtr;              /* Remote transmission request. */
     uint32_t timestamp;   /* Receive timestamp. */
     union {
